Updated some PDUs to reflect published DIS 7 standard. Still need to bounce the full XML standard against the XML file.
git-svn-id: svn://svn.code.sf.net/p/open-dis/code@844 1da32bc5-2e4a-4c7f-92e1-9780f3f753aa

+++ languages/cpp/trunk/cpp/DIS7/VariableParameter.cpp
@@ -5,7 +5,7 @@ using namespace DIS;
 
 VariableParameter::VariableParameter():
    _recordType(0), 
-   _variableParameterFields1(0), 
+   _variableParameterFields1(0.0), 
    _variableParameterFields2(0), 
    _variableParameterFields3(0), 
    _variableParameterFields4(0)
@@ -26,12 +26,12 @@ void VariableParameter::setRecordType(unsigned char pX)
     _recordType = pX;
 }
 
-long VariableParameter::getVariableParameterFields1() const
+double VariableParameter::getVariableParameterFields1() const
 {
     return _variableParameterFields1;
 }
 
-void VariableParameter::setVariableParameterFields1(long pX)
+void VariableParameter::setVariableParameterFields1(double pX)
 {
     _variableParameterFields1 = pX;
 }
